third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/permessage-deflate-window-bits-expected.txt

Test WebSocket permessage-deflate extension: window bits.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing query: "c2s_max_window_bits=8"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('c2s_max_window_bits=8') != -1 is true
PASS event.data is message
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "c2s_max_window_bits=15"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('c2s_max_window_bits=15') != -1 is true
PASS event.data is message
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "c2s_no_context_takeover"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('c2s_no_context_takeover') != -1 is true
PASS event.data is message
onclose() was called.
PASS closeEvent.wasClean is true
Testing query: "c2s_max_window_bits=8&c2s_no_context_takeover"
PASS ws.extensions.search('permessage-deflate') != -1 is true
PASS ws.extensions.search('c2s_max_window_bits=8') != -1 is true
PASS ws.extensions.search('c2s_no_context_takeover') != -1 is true
PASS event.data is message
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
